requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# LangGraph and AI
langgraph==0.0.62
google-generativeai==0.3.2

# Database and Caching
psycopg2-binary==2.9.9
redis==5.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional: For development and testing
pytest==7.4.3
pytest-asyncio==0.21.1 